+++ components/Post/post.js
@@ -1,35 +1,11 @@
-import {
-    Box,
-    useColorModeValue,
-    Modal,
-    ModalOverlay,
-    ModalContent,
-    ModalHeader,
-    ModalFooter,
-    ModalBody,
-    ModalCloseButton,
-    useDisclosure,
-    Button,
-    Slider,
-    SliderTrack,
-    SliderFilledTrack,
-    SliderThumb,
-    Tag,
-    Image as ChakraImage,
-    Text,
-    Avatar,
-    Divider,
-    Icon,
-    IconButton,
-    Input,
-} from "@chakra-ui/react";
-import { AddIconButton, ChargeOutlineButton, CommentIconButton, ShareIconButton, TagAddButton } from "../UI/IconButton";
-import { ThunderboltOutlined, ThunderboltFilled } from "@ant-design/icons";
+import { Box, useColorModeValue, useDisclosure, Image as ChakraImage, Text, Avatar, Divider, Input } from "@chakra-ui/react";
+import { AddIconButton, ChargeOutlineButton, CommentIconButton, ShareIconButton } from "../UI/IconButton";
 import { useState, useEffect } from "react";
 import { nearStore } from "../../stores/near";
 import { Layout } from "antd";
 import { PurpleButton } from "../UI/Buttons";
-import useCustomToast from "../../hooks/useCustomToast";
+import ChargeModal from "./chargeModal";
+import MemberTag from "./tagmembers";
 import TimeAgo from "timeago-react";
 import SongCard from "../Player/songCard";
 
@@ -156,26 +132,7 @@ function Post({ nft, charge}) {
                     <Box onClick={onOpen}><ChargeOutlineButton/>{" "}{currentCharge}</Box>
                     <Box onClick={comment}><CommentIconButton />0</Box>
                     <Box opacity={0.7}><ShareIconButton />0</Box>
-                    <Box style={styles.tag}>
-                        <TagAddButton/>
-                        <Tag borderRadius='full' mt={1}>
-                            <Avatar
-                            name="Tag member 1"
-                            src='https://bit.ly/dan-abramov'
-                            size="2xs"
-                            />
-                            <Avatar
-                            name="Tag member 1"
-                            src='https://bit.ly/code-beast'
-                            size="2xs"
-                            /> 
-                            <Avatar
-                            name="Tag member 1"
-                            src='https://bit.ly/ryan-florence'
-                            size="2xs"
-                            />
-                        </Tag>    
-                    </Box>
+                    <MemberTag style={styles.tag}/>
                 </Footer>
                 <Footer>
                     { commentBox ? 
@@ -205,141 +162,6 @@ function Post({ nft, charge}) {
         </>
     );
 }
-
-const ChargeModal = ({ nft, state }) => {
-    const [isOpen, onClose] = state;
-    const nearState = nearStore((state) => state);
-    const sliderTrack = useColorModeValue("yellow.400", "yellow.400");
-    const sliderTrackBg = useColorModeValue("yellow.100", "yellow.100");
-    const sliderThumbColor = useColorModeValue("gray.900", "gray.900");
-    const [sliderValue, setSliderValue] = useState(0);
-    const postBg = useColorModeValue("#d182ffda", "#171923");
-    const toast = useCustomToast();
-
-    function updateSlider(e) {
-        setSliderValue(e);
-    }
-
-    async function getTotalCharges(_tokenId, _charge) {
-        // Get the post using the tokenId from supabase
-        const { data, error } = await supabase
-            .from("postnft")
-            .select("id, totalCharged")
-            .eq("id", _tokenId);
-        console.log("data received from supabase", data);
-        const newTotalCharge = data.totalCharged + _charge;
-        // Update the post total charge on the db before setting the charge
-        // on page reload
-        setCharge(_tokenId, newTotalCharge);
-
-        if (error) {
-            toast(
-                "error",
-                "Post could not be fetched from Supabase! Error: " +
-                    error.message,
-                "supaErr",
-            );
-            throw error;
-        } else {
-            console.log(" post successfully updated to Supabase"),
-                "supaSuccess";
-            // redirect back to feed
-        }
-    }
-
-    async function setCharge(_tokenId, _charge) {
-        try {
-            await nearState.cnftContract.set_charge({
-                token_id: _tokenId.toString(),
-                charge: _charge.toString(),
-            });
-            toast("success", "Charged " + _charge + "AEX$", "ChargeIderr");
-        } catch (e) {
-            console.log("set charge failed!", e);
-        }
-    }
-
-    async function chargePost() {
-        const amount = 11;
-        nearState.tokenContract
-            .ft_transfer(
-                {
-                    receiver_id: nft.owner_id,
-                    amount: amount.toString(),
-                    memo:
-                        "Charge :zap: from " +
-                        nearState?.accountId +
-                        " for your AEXpost id." +
-                        nft.token_id,
-                },
-                "300000000000000", // attached GAS (optional)
-                1, // attached deposit in yoctoNEAR (optional)
-            )
-            .catch((e) => {
-                console.log("Charge failed!", e);
-                toast("error", "Charge failed!", "ChargeIderr");
-            });
-        //.then(() => setCharge(nft.tokenId, newAmount));
-        onClose();
-    }
-
-    return (
-        <Modal
-            size="xl"
-            isOpen={isOpen}
-            onClose={() => {
-                setSliderValue(0);
-                onClose();
-            }}
-        >
-            <ModalOverlay />
-            <ModalContent bg={postBg}>
-                <ModalHeader>Reward Post</ModalHeader>
-                <ModalCloseButton />
-                <ModalBody>
-                    <Box className="py-2 flex pr-2">
-                        <Box className="mr-4 text-2xl">
-                            <Icon as={ThunderboltFilled} color="yellow" />
-                        </Box>
-
-                        <Slider
-                            onChange={updateSlider}
-                            size={"lg"}
-                            aria-label="pay-slider"
-                            colorScheme={"yellow"}
-                            defaultValue={0}
-                        >
-                            <SliderTrack bg={sliderTrackBg}>
-                                <SliderFilledTrack bg={sliderTrack} />
-                            </SliderTrack>
-                            <SliderThumb color={sliderThumbColor} boxSize={6}>
-                                <small>{sliderValue}</small>
-                            </SliderThumb>
-                        </Slider>
-                    </Box>
-                </ModalBody>
-
-                <ModalFooter>
-                    <Button
-                        variant="ghost"
-                        mr={3}
-                        onClick={() => {
-                            setSliderValue(0);
-                            onClose();
-                        }}
-                    >
-                        Close
-                    </Button>
-                    <Button colorScheme="blue" onClick={chargePost}>
-                        Confirm
-                    </Button>
-                </ModalFooter>
-            </ModalContent>
-        </Modal>
-    );
-};
-
-
 
 
 export default Post;
